% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bernie.R
\name{geom_bernie}
\alias{geom_bernie}
\title{geom_bernie}
\usage{
geom_bernie(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
Add image of bernie to ggplot
}
