% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bernie.R
\docType{data}
\name{Geombernie}
\alias{Geombernie}
\title{Geombernie}
\format{
An object of class \code{Geombernie} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 9.
}
\usage{
Geombernie
}
\description{
Geombernie
}
\keyword{datasets}
